+++ CHANGELOG.md
@@ -43,11 +43,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed redundant "Back to resource" button from crowdsourcing review page
+- [dev dependency] Removed unused `fork-ts-checker-webpack-plugin` package 
 
 ### Security
 
 - [dev dependency] Updated Jest 26 -> 27
-- 
+- [dev dependency] Updated Webpack 4 -> 5
 
 
 ## [2.0.0](https://github.com/digirati-co-uk/madoc-platform/releases/tag/v2.0.0) - 2021-10-09